% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{latest_tok}
\alias{latest_tok}
\title{Make sure not to have overlapping data from multiple tiktok imports}
\usage{
latest_tok(.x)
}
\arguments{
\item{.x}{table. A table of tiktok data structured with the read_tok function.}
}
\description{
The function will get a tiktok input and update it to the highest value in "vanity metrics", such as likes
comments, shares and other aspects. This will allow the user to create multiple extractions
and simplify the process of curating which ones would overlap.
}
\examples{
latest_tok(dabloons)

}
