% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{read_tok}
\alias{read_tok}
\title{Clean data from Zeeswchimer tiktok download}
\usage{
read_tok(.x)
}
\arguments{
\item{.x}{string. Path to data-set.}
}
\description{
The function will clean and organize columns extracted from the the web-app zeeschwimer.
}
\examples{
read_tok("path/to/file")

}
